# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TECA API",
    "description": "\nAPI del e-commerce de muebles **TECA — Para toda la vida**.\n\nCubre la tienda pública (catálogo, carrito, checkout de invitado o con cuenta),\nel área del cliente (pedidos, direcciones, métodos de pago, devoluciones) y el\npanel de administración (usuarios internos, productos, pedidos, cupones,\nauditoría y finanzas).\n\n### Autenticación\nTodas las rutas protegidas usan **JWT Bearer**:\n\n1. `POST /auth/login` con correo y contraseña.\n2. Copia el `access_token` de la respuesta.\n3. Pulsa **Authorize** (arriba a la derecha) y pégalo.\n\n### Roles\n| Rol | Alcance |\n|---|---|\n| `cliente` | Tienda: carrito, pedidos propios, cuenta, reseñas |\n| `admin` | Acceso total al panel |\n| `editor` | Productos |\n| `encargado` | Productos, pedidos y devoluciones |\n| `vendedor` | Pedidos (y lectura de productos) |\n| `soporte` | Devoluciones y lectura de pedidos |\n| `finanzas` | Reportes financieros |\n\n### Convenciones\n- Moneda: balboas (B/.), equivalente a USD.\n- Los identificadores son `ObjectId` de Mongo expuestos como `id` (string).\n- Errores en formato `{\"detail\": \"mensaje\"}` con el código HTTP correspondiente.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Listar productos del catálogo",
        "description": "Alimenta la pantalla **Catálogo de productos** con todos sus filtros de la barra lateral. Solo devuelve productos activos.\n\n**Filtros disponibles**\n- `search`: búsqueda de texto en nombre y descripción (la lupa del header).\n- `category` y `material`: aceptan **varios valores separados por coma** (`?category=sofa,silla`), tal como los checkboxes del filtro lateral.\n- `min_price` / `max_price`: el deslizador de precio.\n\n**Ordenamiento** (`sort`), igual que el desplegable *Ordenar por*:\n| Valor | Significado |\n|---|---|\n| `price_asc` | Precio de menor a mayor |\n| `price_desc` | Precio de mayor a menor |\n| `newest` | Más nuevos |\n| `relevance` | Más relevantes (mejor calificados) — por defecto |\n\n**Respuesta:** `items`, `total`, `page`, `page_size` y `total_pages`, para pintar el texto *«Mostrando 1-8 de 24 productos»* y la paginación.",
        "operationId": "list_products_products_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Búsqueda por texto en nombre y descripción",
              "title": "Search"
            },
            "description": "Búsqueda por texto en nombre y descripción"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Categorías separadas por coma: sofa, silla, mesa, cama, lampara, espejo",
              "title": "Category"
            },
            "description": "Categorías separadas por coma: sofa, silla, mesa, cama, lampara, espejo"
          },
          {
            "name": "material",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Materiales separados por coma: madera, tela, metal, vidrio",
              "title": "Material"
            },
            "description": "Materiales separados por coma: madera, tela, metal, vidrio"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Precio mínimo en balboas (B/.)",
              "title": "Min Price"
            },
            "description": "Precio mínimo en balboas (B/.)"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Precio máximo en balboas (B/.)",
              "title": "Max Price"
            },
            "description": "Precio máximo en balboas (B/.)"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "price_asc",
                "price_desc",
                "newest",
                "relevance"
              ],
              "type": "string",
              "description": "Orden: price_asc (menor a mayor), price_desc (mayor a menor), newest (más nuevos), relevance (más relevantes)",
              "default": "relevance",
              "title": "Sort"
            },
            "description": "Orden: price_asc (menor a mayor), price_desc (mayor a menor), newest (más nuevos), relevance (más relevantes)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página, empezando en 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página, empezando en 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Productos por página (máx. 50)",
              "default": 8,
              "title": "Page Size"
            },
            "description": "Productos por página (máx. 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Detalle de un producto",
        "description": "Datos completos para la ficha del producto: precio, descripción, material, color, stock, galería de imágenes (`images`) y las viñetas de *Detalles del producto* (`dimensions`, `structure`, `warranty`), además de `rating_avg` y `rating_count`.\n\nEl campo `stock` es el que limita el selector de cantidad (*«No puedes seleccionar más de N unidades»*).\n\n**Errores:** `404` si no existe o está inactivo.",
        "operationId": "get_product_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto devuelto por el catálogo",
              "title": "Product Id"
            },
            "description": "ID del producto devuelto por el catálogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/reviews": {
      "get": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Reseñas de un producto",
        "description": "Bloque **Reseñas del producto**. Devuelve:\n\n- `items`: reseñas paginadas (más recientes primero), cada una con nombre del autor, comentario, fecha y `verified_purchase` para la etiqueta *Compra verificada*.\n- `average`: promedio general (el número grande, ej. `4.2`).\n- `distribution`: cantidad de reseñas por estrella (`{\"5\": 22, \"4\": 10, ...}`) para dibujar las barras 5★–1★.\n- `total`: total de reseñas (*«Ver todas las reseñas (11)»*).",
        "operationId": "list_reviews_products__product_id__reviews_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto",
              "title": "Product Id"
            },
            "description": "ID del producto"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página, empezando en 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página, empezando en 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Reseñas por página (máx. 50)",
              "default": 10,
              "title": "Page Size"
            },
            "description": "Reseñas por página (máx. 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Escribir una reseña",
        "description": "Botón *Escribir mi reseña*. Recibe `rating` (1 a 5) y `comment` (máx. 500 caracteres).\n\nLa etiqueta **Compra verificada** se calcula sola: se marca `verified_purchase` en `true` si el usuario tiene un pedido **entregado** que incluye ese producto. Tras guardar, recalcula el promedio y el conteo del producto.\n\n🔒 Requiere token. **Errores:** `404` producto inexistente · `409` si ya reseñaste este producto (una reseña por usuario).",
        "operationId": "create_review_products__product_id__reviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto a reseñar",
              "title": "Product Id"
            },
            "description": "ID del producto a reseñar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reseña"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "409": {
            "description": "Ya publicaste una reseña de este producto"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/products": {
      "get": {
        "tags": [
          "Administración"
        ],
        "summary": "Listar productos (panel)",
        "description": "Tabla de la pantalla **Productos** del panel: nombre, categoría, precio y stock, con el pie *«Mostrando 6 de 12 productos»* y paginación.\n\nA diferencia del catálogo público, **incluye los productos inactivos**.\n\n**Filtros:** `search` (buscador por nombre), `category` y `low_stock=true` para ver solo lo que está por agotarse (stock < 5, la etiqueta *Bajo*).\n\n🔒 Requiere cualquier rol interno.",
        "operationId": "list_products_admin_admin_products_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Buscar por nombre del producto",
              "title": "Search"
            },
            "description": "Buscar por nombre del producto"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Categoría: sofa, silla, mesa, cama, lampara o espejo",
              "title": "Category"
            },
            "description": "Categoría: sofa, silla, mesa, cama, lampara o espejo"
          },
          {
            "name": "low_stock",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Solo productos con stock bajo (< 5)",
              "default": false,
              "title": "Low Stock"
            },
            "description": "Solo productos con stock bajo (< 5)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página, empezando en 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página, empezando en 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Productos por página (máx. 50)",
              "default": 6,
              "title": "Page Size"
            },
            "description": "Productos por página (máx. 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administración"
        ],
        "summary": "Crear un producto",
        "description": "Botón **+ Nuevo** de la pantalla Productos.\n\nCampos obligatorios: `name`, `price` (mayor que 0), `category` y `material`. Opcionales: descripción, color, stock, `images` (URLs de la galería) y las viñetas de la ficha (`dimensions`, `structure`, `warranty`).\n\nNace con calificación en 0 y, si `active` es `true`, aparece de inmediato en el catálogo público. Queda registrado en la auditoría.\n\n🔒 Roles: **admin**, **editor**, **encargado**.",
        "operationId": "create_product_admin_products_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto (crear)"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/admin/products/{product_id}": {
      "patch": {
        "tags": [
          "Administración"
        ],
        "summary": "Editar un producto",
        "description": "Icono de lápiz de la tabla. Actualización **parcial**: envía solo los campos que cambian.\n\nEs también la vía para reponer inventario (`{\"stock\": 20}`) o para ocultar un producto del catálogo sin borrarlo (`{\"active\": false}`).\n\nQueda registrado en la auditoría.\n\n🔒 Roles: **admin**, **editor**, **encargado**. **Errores:** `404` si no existe.",
        "operationId": "update_product_admin_products__product_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto a editar",
              "title": "Product Id"
            },
            "description": "ID del producto a editar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto (editar)"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Administración"
        ],
        "summary": "Eliminar un producto",
        "description": "Modal *«¿Eliminar este producto? Esta opción no se puede deshacer»*. Responde `204` sin cuerpo.\n\nBorra el producto del catálogo de forma permanente. Los pedidos anteriores no se ven afectados: guardan una copia de los datos del producto. Si solo quieres retirarlo de la tienda, usa `PATCH` con `active: false`.\n\nQueda registrado en la auditoría.\n\n🔒 Roles: **admin**, **editor**, **encargado**. **Errores:** `404` si no existe.",
        "operationId": "delete_product_admin_products__product_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto a eliminar",
              "title": "Product Id"
            },
            "description": "ID del producto a eliminar"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Reseña": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Calificación",
            "description": "Estrellas, de 1 a 5"
          },
          "comment": {
            "type": "string",
            "maxLength": 500,
            "title": "Comentario",
            "description": "Máximo 500 caracteres"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "comment"
        ],
        "title": "Reseña",
        "example": {
          "comment": "El sofá llegó en perfecto estado y la entrega fue rápida.",
          "rating": 5
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}